+++ src/lib/libast/hash/hashlib.h
@@ -75,7 +75,7 @@ typedef struct				/* root local pointers		*/
 #define HASHMINSIZE	(1<<4)		/* min table slots (power of 2)	*/
 #define HASHMEANCHAIN	2		/* def resize mean chain len	*/
 
-#define HASHMOD(t,h)	(h &= (t->size - 1))
+#define HASHMOD(t,h)	(h &= (unsigned long)(t->size - 1))
 #define HASHVAL(x)	((x)&~HASH_FLAGS)
 
 #define HASH(r,n,h)	if (r->local->hash) h = r->namesize ? (*r->local->hash)(n, r->namesize) : (*r->local->hash)(n);\
