+++ src/lib/libcmd/context.c
@@ -16,8 +16,8 @@
 *            Johnothan King <johnothanking@protonmail.com>             *
 *                                                                      *
 ***********************************************************************/
-#define CONTEXT_BLOCK		(1024*1024)
-#define CONTEXT_LINE		(4*1024)
+#define CONTEXT_BLOCK		(1024U*1024U)
+#define CONTEXT_LINE		(4U*1024U)
 
 #define _CONTEXT_LINE_PRIVATE_		\
 	unsigned char	show;		\
@@ -102,11 +102,11 @@ context_line(Context_t* cp)
 		lp->span = 0;
 		free(lp->data);
 	}
-	n = cp->end - cp->cur;
+	n = (size_t)(cp->end - cp->cur);
 	if (s = memchr(cp->cur, '\n', n))
 	{
 		lp->data = cp->cur;
-		n = s - cp->cur + 1;
+		n = (size_t)(s - cp->cur + 1);
 		cp->cur += n;
 		lp->size = n;
 	}
@@ -128,11 +128,11 @@ context_line(Context_t* cp)
 			if ((r = sfread(cp->ip, cp->buf, CONTEXT_BLOCK)) <= 0)
 				return 0;
 			cp->end = cp->buf + r;
-			n = (s = memchr(cp->buf, '\n', r)) ? (s - cp->buf + 1) : r;
+			n = (s = memchr(cp->buf, '\n', (size_t)r)) ? (size_t)(s - cp->buf + 1) : (size_t)r;
 			if ((ssize_t)n > (e - t))
 			{
 				r = t - lp->data;
-				m = r + (s - cp->buf);
+				m = (size_t)(r + (s - cp->buf));
 				m = roundof(m, CONTEXT_LINE);
 				if (!(lp->data = oldof(lp->data, char, m, 0)))
 					return 0;
@@ -142,7 +142,7 @@ context_line(Context_t* cp)
 			t += n;
 			cp->cur = cp->buf + n;
 		} while (!s);
-		lp->size = t - lp->data;
+		lp->size = (size_t)(t - lp->data);
 	}
 	lp->line = ++cp->lineno;
 	return lp;
